Avoid bumping the version unnecessarily.
We were getting a lot of version bumps because of finodlks on operations
before we had ever tried to write the file, generating false conflicts and
retries.  It still worked, and the performance impact probably wouldn't be
noticable unless you were creating and writing many small files back to
back, but it was ugly and tended to clutter up logs while debugging other
issues.

Change-Id: Ie0bf4f3162a1d1ec36cf062f94e716ac06c636bb