Fix interaction with self-heal.
The issue is that self-heal might erase the changelog counts for a file
while we're still writing it.  To deal with this, we maintain a "self-heal
version" on the server that's only incremented when a self-heal xattr op
comes in.  This is passed back and forth just like we already do for the
data version, and we treat mismatches the same way too.  The one other
critical piece is that if there's a mismatch then we know the changelog
count was already set to zero.  Thus, a client detecting a mismatch resets
its changelog/retry counter.  A server detecting a mismatch on a changelog
decrement just drops the decrement.  This is the case where the self-heal
happened after the write but before the asynchronous decrement.  Since we
know in this case that the self-heal already set the changelog counter to
zero, our own decrement is superfluous and would in fact cause a counter
underflow if processed, so dropping it is the right thing to do.

Change-Id: I3f720a65c1b08b1838f4dca120f76d5caf5d9cb3