More merge artifacts.
This version works with the official xdata code.  Unfortunately, that
code is grossly inefficient so hsrepl is now up to 12% slower than
before.  At the same time, the applicability of the various
optimizations in AFR have made it up to 68% faster than before.  Net
result is that AFR alone is now up to 8% faster than AFR+hsrepl on my
basic test case.  Now that some of the techniques proven in hsrepl are
being imitated in AFR (e.g. asynchronous changelog decrements) the
balance is likely to tip even further in that direction.  Its original
purpose fulfilled, this is likely to be the last version of hsrepl as a
separate translator.

Change-Id: I2b3719cfd4b21ac118672a2ce5225e02ae0944d6